improvement(knowledge-upload): ui/ux and file structure
