+++ apps/sim/app/w/knowledge/[id]/[documentId]/components/create-chunk-modal/create-chunk-modal.tsx
@@ -1,0 +1,255 @@
+'use client'
+
+import { useRef, useState } from 'react'
+import { AlertCircle, FileText, Loader2, X } from 'lucide-react'
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from '@/components/ui/alert-dialog'
+import { Button } from '@/components/ui/button'
+import { Checkbox } from '@/components/ui/checkbox'
+import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/components/ui/dialog'
+import { Label } from '@/components/ui/label'
+import { Textarea } from '@/components/ui/textarea'
+import { createLogger } from '@/lib/logs/console-logger'
+import type { ChunkData, DocumentData } from '@/stores/knowledge/store'
+
+const logger = createLogger('CreateChunkModal')
+
+interface CreateChunkModalProps {
+  open: boolean
+  onOpenChange: (open: boolean) => void
+  document: DocumentData | null
+  knowledgeBaseId: string
+  onChunkCreated?: (chunk: ChunkData) => void
+}
+
+export function CreateChunkModal({
+  open,
+  onOpenChange,
+  document,
+  knowledgeBaseId,
+  onChunkCreated,
+}: CreateChunkModalProps) {
+  const [content, setContent] = useState('')
+  const [enabled, setEnabled] = useState(true)
+  const [isCreating, setIsCreating] = useState(false)
+  const [error, setError] = useState<string | null>(null)
+  const [showUnsavedChangesAlert, setShowUnsavedChangesAlert] = useState(false)
+  const isProcessingRef = useRef(false)
+
+  const hasUnsavedChanges = content.trim().length > 0
+
+  const handleCreateChunk = async () => {
+    if (!document || content.trim().length === 0 || isProcessingRef.current) {
+      if (isProcessingRef.current) {
+        logger.warn('Chunk creation already in progress, ignoring duplicate request')
+      }
+      return
+    }
+
+    try {
+      isProcessingRef.current = true
+      setIsCreating(true)
+      setError(null)
+
+      const response = await fetch(
+        `/api/knowledge/${knowledgeBaseId}/documents/${document.id}/chunks`,
+        {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+          },
+          body: JSON.stringify({
+            content: content.trim(),
+            enabled,
+          }),
+        }
+      )
+
+      if (!response.ok) {
+        const result = await response.json()
+        throw new Error(result.error || 'Failed to create chunk')
+      }
+
+      const result = await response.json()
+
+      if (result.success && result.data) {
+        logger.info('Chunk created successfully:', result.data.id)
+
+        if (onChunkCreated) {
+          onChunkCreated(result.data)
+        }
+
+        onClose()
+      } else {
+        throw new Error(result.error || 'Failed to create chunk')
+      }
+    } catch (err) {
+      logger.error('Error creating chunk:', err)
+      setError(err instanceof Error ? err.message : 'An error occurred')
+    } finally {
+      isProcessingRef.current = false
+      setIsCreating(false)
+    }
+  }
+
+  const onClose = () => {
+    onOpenChange(false)
+    // Reset form state when modal closes
+    setContent('')
+    setEnabled(true)
+    setError(null)
+    setShowUnsavedChangesAlert(false)
+  }
+
+  const handleCloseAttempt = () => {
+    if (hasUnsavedChanges && !isCreating) {
+      setShowUnsavedChangesAlert(true)
+    } else {
+      onClose()
+    }
+  }
+
+  const handleConfirmDiscard = () => {
+    setShowUnsavedChangesAlert(false)
+    onClose()
+  }
+
+  const isFormValid = content.trim().length > 0 && content.trim().length <= 10000
+
+  return (
+    <>
+      <Dialog open={open} onOpenChange={handleCloseAttempt}>
+        <DialogContent
+          className='flex h-[74vh] flex-col gap-0 overflow-hidden p-0 sm:max-w-[600px]'
+          hideCloseButton
+        >
+          <DialogHeader className='flex-shrink-0 border-b px-6 py-4'>
+            <div className='flex items-center justify-between'>
+              <DialogTitle className='font-medium text-lg'>Create Chunk</DialogTitle>
+              <Button
+                variant='ghost'
+                size='icon'
+                className='h-8 w-8 p-0'
+                onClick={handleCloseAttempt}
+              >
+                <X className='h-4 w-4' />
+                <span className='sr-only'>Close</span>
+              </Button>
+            </div>
+          </DialogHeader>
+
+          <div className='flex flex-1 flex-col overflow-hidden'>
+            <div className='scrollbar-thin scrollbar-thumb-muted-foreground/20 hover:scrollbar-thumb-muted-foreground/25 scrollbar-track-transparent min-h-0 flex-1 overflow-y-auto px-6'>
+              <div className='py-4'>
+                <div className='space-y-4'>
+                  {/* Document Info */}
+                  <div className='flex items-center gap-3 rounded-lg border bg-muted/30 p-4'>
+                    <FileText className='h-5 w-5 text-muted-foreground' />
+                    <div className='min-w-0 flex-1'>
+                      <p className='font-medium text-sm'>
+                        {document?.filename || 'Unknown Document'}
+                      </p>
+                      <p className='text-muted-foreground text-xs'>Adding chunk to this document</p>
+                    </div>
+                  </div>
+
+                  {/* Content Input */}
+                  <div className='space-y-2'>
+                    <Label htmlFor='content' className='font-medium text-sm'>
+                      Chunk Content
+                    </Label>
+                    <Textarea
+                      id='content'
+                      value={content}
+                      onChange={(e) => setContent(e.target.value)}
+                      placeholder='Enter the content for this chunk...'
+                      className='min-h-[200px] resize-none'
+                      disabled={isCreating}
+                    />
+                    <div className='flex items-center justify-between text-muted-foreground text-xs'>
+                      <span>{content.length}/10000 characters</span>
+                      {content.length > 10000 && (
+                        <span className='text-red-500'>Content too long</span>
+                      )}
+                    </div>
+                  </div>
+
+                  {/* Enabled Toggle */}
+                  <div className='flex items-center space-x-2'>
+                    <Checkbox
+                      id='enabled'
+                      checked={enabled}
+                      onCheckedChange={(checked) => setEnabled(checked as boolean)}
+                      disabled={isCreating}
+                      className='h-4 w-4'
+                    />
+                    <Label htmlFor='enabled' className='text-sm'>
+                      Enable chunk for search
+                    </Label>
+                  </div>
+
+                  {/* Error Display */}
+                  {error && (
+                    <div className='flex items-center gap-2 rounded-md border border-red-200 bg-red-50 p-3'>
+                      <AlertCircle className='h-4 w-4 text-red-600' />
+                      <p className='text-red-800 text-sm'>{error}</p>
+                    </div>
+                  )}
+                </div>
+              </div>
+            </div>
+
+            {/* Footer */}
+            <div className='mt-auto border-t px-6 pt-4 pb-6'>
+              <div className='flex justify-between'>
+                <Button variant='outline' onClick={handleCloseAttempt} disabled={isCreating}>
+                  Cancel
+                </Button>
+                <Button
+                  onClick={handleCreateChunk}
+                  disabled={!isFormValid || isCreating}
+                  className='bg-[#701FFC] font-[480] text-primary-foreground shadow-[0_0_0_0_#701FFC] transition-all duration-200 hover:bg-[#6518E6] hover:shadow-[0_0_0_4px_rgba(127,47,255,0.15)]'
+                >
+                  {isCreating ? (
+                    <>
+                      <Loader2 className='mr-2 h-4 w-4 animate-spin' />
+                      Creating...
+                    </>
+                  ) : (
+                    'Create Chunk'
+                  )}
+                </Button>
+              </div>
+            </div>
+          </div>
+        </DialogContent>
+      </Dialog>
+
+      {/* Unsaved Changes Alert */}
+      <AlertDialog open={showUnsavedChangesAlert} onOpenChange={setShowUnsavedChangesAlert}>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>Discard changes?</AlertDialogTitle>
+            <AlertDialogDescription>
+              You have unsaved changes. Are you sure you want to close without saving?
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel onClick={() => setShowUnsavedChangesAlert(false)}>
+              Keep editing
+            </AlertDialogCancel>
+            <AlertDialogAction onClick={handleConfirmDiscard}>Discard changes</AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
+    </>
+  )
+}
