+++ apps/sim/app/w/knowledge/[id]/[documentId]/document.tsx
@@ -10,9 +10,10 @@ import { useDocumentChunks } from '@/hooks/use-knowledge'
 import { type ChunkData, type DocumentData, useKnowledgeStore } from '@/stores/knowledge/store'
 import { useSidebarStore } from '@/stores/sidebar/store'
 import { KnowledgeHeader } from '../../components/knowledge-header/knowledge-header'
-import { CreateChunkModal } from './components/create-chunk-modal'
+import { CreateChunkModal } from './components/create-chunk-modal/create-chunk-modal'
+import { DeleteChunkModal } from './components/delete-chunk-modal/delete-chunk-modal'
 import { DocumentLoading } from './components/document-loading'
-import { EditChunkModal } from './components/edit-chunk-modal'
+import { EditChunkModal } from './components/edit-chunk-modal/edit-chunk-modal'
 
 const logger = createLogger('Document')
 
@@ -51,6 +52,8 @@ export function Document({
   const [selectedChunk, setSelectedChunk] = useState<ChunkData | null>(null)
   const [isModalOpen, setIsModalOpen] = useState(false)
   const [isCreateChunkModalOpen, setIsCreateChunkModalOpen] = useState(false)
+  const [chunkToDelete, setChunkToDelete] = useState<ChunkData | null>(null)
+  const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false)
 
   const [document, setDocument] = useState<DocumentData | null>(null)
   const [isLoadingDocument, setIsLoadingDocument] = useState(true)
@@ -170,32 +173,28 @@ export function Document({
     }
   }
 
-  const handleDeleteChunk = async (chunkId: string) => {
-    try {
-      const response = await fetch(
-        `/api/knowledge/${knowledgeBaseId}/documents/${documentId}/chunks/${chunkId}`,
-        {
-          method: 'DELETE',
-        }
-      )
-
-      if (!response.ok) {
-        throw new Error('Failed to delete chunk')
-      }
-
-      const result = await response.json()
-
-      if (result.success) {
-        await refreshChunks()
-        setSelectedChunks((prev) => {
-          const newSet = new Set(prev)
-          newSet.delete(chunkId)
-          return newSet
-        })
-      }
-    } catch (err) {
-      logger.error('Error deleting chunk:', err)
+  const handleDeleteChunk = (chunkId: string) => {
+    const chunk = chunks.find((c) => c.id === chunkId)
+    if (chunk) {
+      setChunkToDelete(chunk)
+      setIsDeleteModalOpen(true)
     }
+  }
+
+  const handleChunkDeleted = async () => {
+    await refreshChunks()
+    if (chunkToDelete) {
+      setSelectedChunks((prev) => {
+        const newSet = new Set(prev)
+        newSet.delete(chunkToDelete.id)
+        return newSet
+      })
+    }
+  }
+
+  const handleCloseDeleteModal = () => {
+    setIsDeleteModalOpen(false)
+    setChunkToDelete(null)
   }
 
   const handleSelectChunk = (chunkId: string, checked: boolean) => {
@@ -593,6 +592,16 @@ export function Document({
         document={document}
         knowledgeBaseId={knowledgeBaseId}
         onChunkCreated={handleChunkCreated}
+      />
+
+      {/* Delete Chunk Modal */}
+      <DeleteChunkModal
+        chunk={chunkToDelete}
+        knowledgeBaseId={knowledgeBaseId}
+        documentId={documentId}
+        isOpen={isDeleteModalOpen}
+        onClose={handleCloseDeleteModal}
+        onChunkDeleted={handleChunkDeleted}
       />
     </div>
   )
